+++ vcx/libvcx/src/v3/messages/mime_type.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
+pub enum MimeType {
+    #[serde(rename = "text/plain")]
+    Plain
+}